Work on getting testing in place and express updated.
