+++ .dockerignore
@@ -1,3 +1,4 @@
 node_modules
 .git
 Dockerfile
+yarn.lock
