+++ .gitignore
@@ -7,3 +7,4 @@ public/js/app.js
 public/js/client.js
 public/js/libs.js
 public/js/idbmap.js
+yarn.lock
