+++ postinstall.sh
@@ -10,10 +10,7 @@ node_modules/bower/bin/bower install --allow-root
 #compile jsx to build files for server side use of React
 babel public/client/js/react/src/ --out-dir public/client/js/react/build/ --blacklist strict
 #compile full client side file with jsx transforms for browser side client
-#for Babel 5.8.34
-browserify -o public/js/client.js  public/client/js/main.js -t [ babelify --presets [ es2015 react ] ]
-#for Babel 5.8.20
-#node_modules/browserify/bin/cmd.js -o public/js/client.js  public/client/js/main.js -t [ babelify ]
+browserify -o public/js/client.js  public/client/js/main.js -t [ babelify ]
 #minify client side file
 uglifyjs -o public/js/client.js public/js/client.js
